Consolidate mwoauth handshake into auth_service
Move MediaWiki OAuth handshake helpers into auth_service.py: get_handshaker, start_login, complete_login, and OAuthIdentityError. Remove the now-redundant mwoauth_handshake.py and its tests, update imports in auth routes and integration/unit tests to use auth_service, and adjust test fixtures to use types.SimpleNamespace where needed. This centralizes OAuth logic and cleans up duplicate modules.
