+++ tests/integration/public/auth/test_oauth_helpers_integration.py
@@ -4,7 +4,7 @@ from types import SimpleNamespace
 
 from src.main_app import AppFactory
 from src.main_app.config import TestingConfig
-from src.main_app.shared.auth import mwoauth_handshake as oauth_helpers
+from src.main_app.shared.auth import auth_service as oauth_helpers
 
 
 class StubConsumerToken:
