+++ tests/unit/public/auth/test_auth_utils.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+from types import SimpleNamespace
 from unittest.mock import MagicMock
 
 import pytest
@@ -12,10 +13,9 @@ from src.main_app.public.auth import utils as auth_utils
 
 @pytest.fixture(autouse=True)
 def mock_settings(monkeypatch: pytest.MonkeyPatch) -> None:
-    import types
 
-    fake_settings = types.SimpleNamespace(
-        cookie=types.SimpleNamespace(name="auth_cookie"),
+    fake_settings = SimpleNamespace(
+        cookie=SimpleNamespace(name="auth_cookie"),
     )
     monkeypatch.setattr("src.main_app.public.auth.utils.settings", fake_settings)
